# apps/api/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: "Afrik'Fid API"
  version: "1.0.0"
  description: |
    Passerelle de paiement B2B multi-pays avec fidélisation intégrée pour l'Afrique francophone et anglophone.

    ## Authentification

    L'API supporte deux mécanismes d'authentification :

    **1. JWT Bearer** — Pour le tableau de bord admin et marchand
    ```
    Authorization: Bearer <accessToken>
    ```
    Obtenir un token via `POST /auth/admin/login` ou `POST /auth/merchant/login`.

    **2. API Key** — Pour les intégrations marchand (paiements, confirmations, statuts)
    ```
    X-API-Key: af_pub_xxxxxxxxxxxx
    ```
    Ajouter `X-Sandbox: true` pour utiliser les clés sandbox et simuler les paiements.

    ## Modèle de fidélité X/Y/Z

    Chaque paiement est réparti entre trois parties :
    - **X%** — Commission marchand (remise accordée au client ou gardée)
    - **Y%** — Rebate client (selon statut loyalty : OPEN=0%, LIVE=5%, GOLD=8%, ROYAL=10%)
    - **Z%** — Commission plateforme Afrik'Fid

    ## Environnements

    | Environnement | Base URL | Clé |
    |---|---|---|
    | Production | `https://api.afrikfid.com/api/v1` | `af_pub_xxx` |
    | Sandbox | `https://api.afrikfid.com/api/v1` + `X-Sandbox: true` | `af_sandbox_pub_xxx` |
    | Local | `http://localhost:4001/api/v1` | Clé de test |

  contact:
    name: "Support Afrik'Fid"
    email: support@afrikfid.com
  license:
    name: Propriétaire
servers:
  - url: http://localhost:4001/api/v1
...
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via /auth/admin/login ou /auth/merchant/login
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API publique du marchand (af_pub_xxx en prod, af_sandbox_pub_xxx en sandbox)
    HmacSignature:
      type: apiKey
      in: header
      name: X-AfrikFid-Signature
      description: |
        Signature HMAC-SHA256 optionnelle pour l'intégrité des requêtes (CDC §4.5.2).
        Format : `sha256=<hex>`
        Calcul : `HMAC-SHA256(api_key_secret, raw_request_body)`
        En mode `REQUIRE_HMAC_SIGNATURE=true`, ce header devient obligatoire sur `/payments/initiate` et `/payments/{id}/refund`.

  # ─── Schemas ───────────────────────────────────────────────────────────────
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: "Données invalides"
      required: [error]
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
...
    PaymentLink:
      type: object
      properties:
        id: { type: string, format: uuid }
        merchantId: { type: string, format: uuid }
        code: { type: string, example: "PL-ABCD-1234" }
        amount: { type: number, nullable: true, example: 5000 }
        currency: { type: string, example: XOF }
        description: { type: string, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
        maxUses: { type: integer, example: 1 }
        usesCount: { type: integer, example: 0 }
        status: { type: string, enum: [active, used, cancelled, expired] }
        createdAt: { type: string, format: date-time }
...
  responses:
    Unauthorized:
      description: Token manquant ou invalide
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example: { error: "TOKEN_INVALID", message: "Token invalide" }
...
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Données invalides
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationError' }

# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /payment-links:
    post:
      tags: [Payment Links]
      summary: Créer un lien de paiement
      description: Crée un lien partageable permettant à un client de payer sans intégration technique.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  description: Montant fixe (omis = montant libre, le client saisit)
                  example: 5000
                currency: { type: string, example: XOF }
                description: { type: string, example: "Prestation conseil" }
                expires_in_hours:
                  type: integer
                  description: Durée de validité en heures (défaut 24h)
                  example: 48
                max_uses:
                  type: integer
                  description: Nombre d'utilisations max (défaut 1)
      responses:
        '201':
          description: Lien créé
          content:
            application/json:
              schema:
                type: object
                properties:
                  link: { $ref: '#/components/schemas/PaymentLink' }
                  payUrl:
                    type: string
                    format: uri
                    description: URL publique à partager
                    example: "https://afrikfid.com/pay/PL-ABCD-1234"
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags: [Payment Links]
      summary: Lister les liens du marchand connecté
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des liens
          content:
            application/json:
              schema:
                type: object
                properties:
                  links:
                    type: array
                    items: { $ref: '#/components/schemas/PaymentLink' }
  /payment-links/{id}:
    delete:
      tags: [Payment Links]
      summary: Annuler un lien de paiement
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Lien annulé
        '404':
          $ref: '#/components/responses/NotFound'
  /payment-links/{code}/info:
    get:
      tags: [Payment Links]
      summary: Informations publiques d'un lien (page de paiement)
      parameters:
        - in: path
          name: code
          required: true
          schema: { type: string, example: "PL-ABCD-1234" }
      responses:
        '200':
          description: Détails du lien
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: string }
                  merchantName: { type: string }
                  amount: { type: number, nullable: true }
                  currency: { type: string }
                  description: { type: string, nullable: true }
                  expiresAt: { type: string, format: date-time }
                  rebateMode: { type: string }
                  operators: { type: array, items: { type: string } }
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Lien expiré ou épuisé
  /payment-links/{code}/identify-client:
    post:
      tags: [Payment Links]
      summary: Identifier un client avant paiement
      parameters:
        - in: path
          name: code
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone: { type: string }
                afrikfid_id: { type: string }
      responses:
        '200':
          description: Client trouvé ou non
          content:
            application/json:
              schema:
                type: object
                properties:
                  found: { type: boolean }
                  client:
                    nullable: true
                    type: object
                    properties:
                      afrikfidId: { type: string }
                      fullName: { type: string }
                      loyaltyStatus: { type: string }
                      clientRebatePercent: { type: number }
                      walletBalance: { type: number }
  /payment-links/{code}/pay:
    post:
      tags: [Payment Links]
      summary: Initier un paiement via lien
      parameters:
        - in: path
          name: code
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, payment_operator]
              properties:
                phone: { type: string, example: "+2250700001" }
                payment_operator: { type: string, enum: [ORANGE, MTN, WAVE, AIRTEL, MOOV] }
                afrikfid_id: { type: string, description: "ID client pour créditer la fidélité" }
                custom_amount: { type: number, description: "Montant si lien à montant libre" }
      responses:
        '200':
          description: Paiement initié
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactionId: { type: string }
                  reference: { type: string }
                  distribution: { type: object }
                  message: { type: string }
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── MERCHANTS ─────────────────────────────────────────────────────────────
...